modularized sensor object and 1 unit test for sensorManager

+++ onboard-software/tubular-platformIO/src/_4_heater.ino
@@ -50,7 +50,7 @@ void readingData(void *pvParameters)
   while(1)
   {
     //Reads the current mode
-    static int currMode = getMode();
+    static uint8_t currMode = getMode();
     //Reads the temperature at the two sensors
     tempAtHtr  = readData(0);
  //   tempAtHtr[1]  = *readData(0);//Must make sure to get the correct tempSensors and add pointer.
